Simple viridisLite wrapper

Simple viridisLite wrapper

viridisPalette(
  total_colors,
  color_palette = "viridis",
  alpha = 1,
  direction = 1
)

Arguments

total_colors

number of colors desired

color_palette

A character string indicating the color map option to use. Eight options are available: "viridis","cividis","magma","inferno","plasma","rocket","mako","turbo"

alpha

The alpha transparency, a number in [0,1]

direction

Sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest. If -1, the order of colors is reversed

Value

Specified color palette of length `total_colors`

Examples

viridisPalette(5)
#> [1] "#440154FF" "#3B528BFF" "#21908CFF" "#5DC863FF" "#FDE725FF"
viridisPalette(5,color_palette='magma',alpha=0.5)
#> [1] "#00000480" "#51127C80" "#B6367980" "#FB886180" "#FCFDBF80"
viridisPalette(5,color_palette='plasma',alpha=0.6,direction=-1)
#> [1] "#F0F92199" "#F8944199" "#CC467899" "#7E03A899" "#0D088799"